Move MessageCheckBox widget from workers/, to helperwidgets.py
